# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_logo_tournaments__tournament_id__logo_post": {
        "properties": {
          "file": {
            "anyOf": [
              {
                "format": "binary",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_logo_tournaments__tournament_id__logo_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SuccessResponse": {
        "properties": {
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "SuccessResponse",
        "type": "object"
      },
...
      "Tournament": {
        "properties": {
          "auto_assign_courts": {
            "title": "Auto Assign Courts",
            "type": "boolean"
          },
          "club_id": {
            "title": "Club Id",
            "type": "integer"
          },
          "created": {
            "format": "date-time",
            "title": "Created",
            "type": "string"
          },
          "dashboard_endpoint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dashboard Endpoint"
          },
          "dashboard_public": {
            "title": "Dashboard Public",
            "type": "boolean"
          },
          "duration_minutes": {
            "minimum": 1.0,
            "title": "Duration Minutes",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "logo_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Path"
          },
          "margin_minutes": {
            "minimum": 0.0,
            "title": "Margin Minutes",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "players_can_be_in_multiple_teams": {
            "title": "Players Can Be In Multiple Teams",
            "type": "boolean"
          },
          "start_time": {
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/TournamentStatus",
            "default": "OPEN"
          }
        },
        "required": [
          "club_id",
          "name",
          "created",
          "start_time",
          "duration_minutes",
          "margin_minutes",
          "dashboard_public",
          "dashboard_endpoint",
          "logo_path",
          "players_can_be_in_multiple_teams",
          "auto_assign_courts",
          "status",
          "id"
        ],
        "title": "Tournament",
        "type": "object"
      },
      "TournamentBody": {
        "properties": {
          "auto_assign_courts": {
            "title": "Auto Assign Courts",
            "type": "boolean"
          },
          "club_id": {
            "title": "Club Id",
            "type": "integer"
          },
          "dashboard_endpoint": {
            "anyOf": [
              {},
              {
                "type": "string"
              }
            ],
            "title": "Dashboard Endpoint"
          },
          "dashboard_public": {
            "title": "Dashboard Public",
            "type": "boolean"
          },
          "duration_minutes": {
            "minimum": 1.0,
            "title": "Duration Minutes",
            "type": "integer"
          },
          "margin_minutes": {
            "minimum": 0.0,
            "title": "Margin Minutes",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "players_can_be_in_multiple_teams": {
            "title": "Players Can Be In Multiple Teams",
            "type": "boolean"
          },
          "start_time": {
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          }
        },
        "required": [
          "start_time",
          "name",
          "dashboard_public",
          "dashboard_endpoint",
          "players_can_be_in_multiple_teams",
          "auto_assign_courts",
          "duration_minutes",
          "margin_minutes",
          "club_id"
        ],
        "title": "TournamentBody",
        "type": "object"
      },
...
      "TournamentChangeStatusBody": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/TournamentStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "TournamentChangeStatusBody",
        "type": "object"
      },
      "TournamentResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Tournament"
          }
        },
        "required": [
          "data"
        ],
        "title": "TournamentResponse",
        "type": "object"
      },
      "TournamentStatus": {
        "enum": [
          "OPEN",
          "ARCHIVED"
        ],
        "title": "TournamentStatus",
        "type": "string"
      },
      "TournamentUpdateBody": {
        "properties": {
          "auto_assign_courts": {
            "title": "Auto Assign Courts",
            "type": "boolean"
          },
          "dashboard_endpoint": {
            "anyOf": [
              {},
              {
                "type": "string"
              }
            ],
            "title": "Dashboard Endpoint"
          },
          "dashboard_public": {
            "title": "Dashboard Public",
            "type": "boolean"
          },
          "duration_minutes": {
            "minimum": 1.0,
            "title": "Duration Minutes",
            "type": "integer"
          },
          "margin_minutes": {
            "minimum": 0.0,
            "title": "Margin Minutes",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "players_can_be_in_multiple_teams": {
            "title": "Players Can Be In Multiple Teams",
            "type": "boolean"
          },
          "start_time": {
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          }
        },
        "required": [
          "start_time",
          "name",
          "dashboard_public",
          "dashboard_endpoint",
          "players_can_be_in_multiple_teams",
          "auto_assign_courts",
          "duration_minutes",
          "margin_minutes"
        ],
        "title": "TournamentUpdateBody",
        "type": "object"
      },
      "TournamentsResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Tournament"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "TournamentsResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "\n### Description\nThis API allows you to do everything the frontend of [Bracket](https://github.com/evroon/bracket)\nallows you to do (the frontend uses this API as well).\n\nFore more information, see the [documentation](https://docs.bracketapp.nl).\n\n### Table of Contents\n*(links only work for [ReDoc](https://api.bracketapp.nl/redoc), not for Swagger UI)*\n\n- [Auth](#tag/Auth)\n\n- [Clubs](#tag/Clubs)\n\n- [Courts](#tag/Courts)\n\n- [Officials](#tag/Officials)\n\n- [Official Portal](#tag/Official-Portal)\n\n- [Tournament Breaks](#tag/Tournament-Breaks)\n\n- [Internals](#tag/Internals)\n\n- [Matches](#tag/Matches)\n\n- [Players](#tag/Players)\n\n- [Rankings](#tag/Rankings)\n\n- [Rounds](#tag/Rounds)\n\n- [Stage Items](#tag/Stage-Items)\n\n- [Stage Item Inputs](#tag/Stage-Item-Inputs)\n\n- [Stages](#tag/Stages)\n\n- [Teams](#tag/Teams)\n\n- [Tournaments](#tag/Tournaments)\n\n- [Users](#tag/Users)\n\n### Links\nGitHub: <https://github.com/evroon/bracket>\n\nDocs: <https://docs.bracketapp.nl>\n\nDemo: <https://www.bracketapp.nl/demo>\n\nAPI docs (Redoc): <https://api.bracketapp.nl/redoc>\n\nAPI docs (Swagger UI): <https://api.bracketapp.nl/docs>\n",
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
    },
    "summary": "API for Bracket, an open source tournament system.",
    "title": "Bracket API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/tournaments": {
      "get": {
        "operationId": "get_tournaments_tournaments_get",
        "parameters": [
          {
            "in": "query",
            "name": "filter_",
            "required": false,
            "schema": {
              "default": "OPEN",
              "enum": [
                "ALL",
                "OPEN",
                "ARCHIVED"
              ],
              "title": "Filter ",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "endpoint_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Endpoint Name"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Tournaments",
        "tags": [
          "Tournaments"
        ]
      },
      "post": {
        "operationId": "create_tournament_tournaments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TournamentBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Tournament",
        "tags": [
          "Tournaments"
        ]
      }
    },
    "/tournaments/{tournament_id}": {
      "delete": {
        "operationId": "delete_tournament_tournaments__tournament_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Tournament",
        "tags": [
          "Tournaments"
        ]
      },
      "get": {
        "operationId": "get_tournament_tournaments__tournament_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Tournament",
        "tags": [
          "Tournaments"
        ]
      },
      "put": {
        "operationId": "update_tournament_by_id_tournaments__tournament_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TournamentUpdateBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Tournament By Id",
        "tags": [
          "Tournaments"
        ]
      }
    },
...
    "/tournaments/{tournament_id}/change-status": {
      "post": {
        "description": "Make a tournament archived or non-archived.",
        "operationId": "change_status_tournaments__tournament_id__change_status_post",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TournamentChangeStatusBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Change Status",
        "tags": [
          "Tournaments"
        ]
      }
    },
...
    "/tournaments/{tournament_id}/logo": {
      "get": {
        "operationId": "get_tournament_logo_tournaments__tournament_id__logo_get",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Tournament Logo",
        "tags": [
          "Tournaments"
        ]
      },
      "post": {
        "operationId": "upload_logo_tournaments__tournament_id__logo_post",
        "parameters": [
          {
            "in": "path",
            "name": "tournament_id",
            "required": true,
            "schema": {
              "title": "Tournament Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_logo_tournaments__tournament_id__logo_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TournamentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Upload Logo",
        "tags": [
          "Tournaments"
        ]
      }
    },
...
  }
}